feat(address-service): DOMA-5129 Form types was moved to more specific place
