+++ apps/condo/domains/property/components/BasePropertyForm/index.tsx
@@ -5,13 +5,13 @@ import { useLayoutContext } from '@condo/domains/common/components/LayoutContext
 import Prompt from '@condo/domains/common/components/Prompt'
 import { useValidations } from '@condo/domains/common/hooks/useValidations'
 import { AddressSuggestionsSearchInput } from '@condo/domains/property/components/AddressSuggestionsSearchInput'
-import { TSelectedAddressSuggestion } from '@condo/domains/property/components/types'
+import { TSelectedAddressSuggestion } from '@condo/domains/property/components/BasePropertyForm/types'
 import { usePropertyValidations } from '@condo/domains/property/hooks/usePropertyValidations'
 import { IPropertyFormState } from '@condo/domains/property/utils/clientSchema/Property'
 import { omitRecursively } from '@open-condo/keystone/fields/Json/utils/cleaner'
 import { useIntl } from '@open-condo/next/intl'
-import { Col, Form, FormInstance, notification, Row, RowProps } from 'antd'
 import { Typography } from '@open-condo/ui/dist'
+import { Col, Form, FormInstance, notification, Row, RowProps } from 'antd'
 import dayjs from 'dayjs'
 import get from 'lodash/get'
 import isEmpty from 'lodash/isEmpty'
